+++ orchestrator/work/worker.go
@@ -157,8 +157,8 @@ func (w *RemoteWorker) Work(ctx context.Context, unit stage.Unit, startBlock uin
 					return err
 				}
 
-				if streamOutput {
-					// never retry for jobs that stream blocks, we don't know how much data they already sent
+				if streamOutput && upstream.DataSent() {
+					// never retry for jobs that stream blocks and have already sent some data
 					segmentStart := request.SegmentNumber * request.SegmentSize
 					return derr.NewFatalError(fmt.Errorf("segment [%d-%d] failed while streaming data: %w", segmentStart, segmentStart+request.SegmentSize, err))
 				}